fix Ph value to double

+++ src/ru/proletov/ckdmbd/models/Ph.java
@@ -2,7 +2,7 @@ package ru.proletov.ckdmbd.models;
 
 public class Ph extends LaboratoryInvestigation {
 
-    public Ph(final float value) {
+    public Ph(final double value) {
         super(value);
     }
